ParseXS: standard_typemap_locations(): skip -e
standard_typemap_locations() returns a set of filenames for where a
typemap file map be expected, relative to both the current directory and
to @INC.

The @INC files are filtered using -e, but the others (../../typemap etc)
aren't.

This commit removes the -e filtering for @INC, for consistency.
This means that standard_typemap_locations() now returns a list of
filenames of *all* the standard locations where a typemap file *might*
be found, regardless of whether the file exists. Previously it was a
weird hybrid.

This commit should make no difference in practice, since the one caller
of this function, process_typemaps() does it's own -f filtering too.

So really the net effect of this commit is to remove a few duplicate
stat() OS calls.

It also allows one existing test to be simplified, since now there will
always be one returned entry per @INC entry, whereas before it might
have been less.