mktables: Clean up hieroglyphs database
Some of the properties in this Unikemet d.b. are provisional, so Unicode
doesn't furnish information about them.  So we have to do it ourselves.
I didn't get it right previously.

This also changes one property to an enum, which it should have been all
along, and adds the current possible enum values.