Fix \b{gcb} bug for Indic Conjunct Break
Its too complicated to go into here, but some Asian languages have a
different method than Western ones for determining grapheme clusters.
I'll abbreviate, and not bother with the long names that don't mean
anything to a non-specialist.  But the regular expression that is
supposed to match is

    qr/ C [EL]* L [EL]* C /x

This is matched with a little DFA inside regexec.c.  I rewrote that to

    qr/ C [EL]* C /x

and kept a count of the L's encountered.  If at least one L occurred, it
was a match.

The problem is that I wrote a do { } while(), and it really needed a
plain while(), so it was getting false positives in some circumstances.

It passed the extensive tests furnished by Unicode for 16.0.  They have
provided a new test file for 17.0, which has new tests, and this failed
for a Balinese example.

I also changed from using a counter, where a simple bool will do.

This fix applies to 16.0 as well as 17.0.