test.pl: Don't use watchdog thread unless safe signals
In the thread documentation, I noticed that unsafe signals preclude
killing threads with a signal.  But this is the normal method for
clearing a watchdog timer under threads.  I doubt that people are
compiling with PERL_OLD_SIGNALS these days, so I didn't check for that,
but its easy enough to check for the environment variable that does the
same thing at runtime.