test.pl: Use END blocks to cancel watchdog timers
In doing follow up work to a7f3f23c846f4fd664a718235fdb2ad97992c26e
"Turn off watchdog when done in tests", I realized that my solution was
suboptimal, and that the code already existed in test.pl to do things
better.

What that code does is to create an END block to cancel the watchdog
upon program exit.  I think that is a better solution than to force
the addition of explicit calls to watchdog(0).  This commit creates a
standard way to specify the code that does the cancellation, and to use
it, not only when the END block does, but also when the timer is
explicitly cancelled.  That standard method is to create a string that
basically gets evaled.   This entailed making any watchdog thread object
into a global so that a reference to it could be stringified for the
eval.

This means the calls to watchdog(0) that occur at the end of the file
that were added in a7f3f23c846f4fd664a718235fdb2ad97992c26e aren't
necessary.  The END block takes care of it.  The reason to keep them
would be if we want to add a porting test that every watchdog is cleared.