unTODO test in GH 21827 TODO tests
This commit restructures the TODO tests for GH 21827 so that the test
that passes is not marked as TODO while the failing tests still are.
warnings_like() had to be swapped out for capture_warnings() with
manual testing so that we could have granular control over the
tests, as we did not have such granular control with the three tests
that were originally generated by warnings_like().