Improve -DNO_SHORT_NAMES handling
This option has been broken for a long time, but this commit goes a long
way to fixing that.

It works by creating a new file long_names.c that is automatically
generated by regen/embed.pl.  This file contains the API elements that
normally are macros that don't have a separate long name.  when
NO_SHORT_NAMES is defined, those macros are suppressed, and instead the
long-name functions in this file get compiled

The next yak (maybe the only one) needing to be shaved after this is
inline functions that use short names.  Since they currently are
included in the user space, those short names need to be defined, but
we'rev compiling without short names, so it doesn't work.

We could require inline functions to use long names, but I don't like
that, and I suspect you don't either.

To me, the easiest other option is to just not have inline functions
outside the core when short names are forbidden.  We could easily
convert them from being inline, and only include them from long_names.c.
Thus they would become like any other function in the file.