test.pl: Don't assume existing watchdog uses alarm()
The fallback method of setting a watchdog timer is to use alarm().
Prior to this commit, the code presumed that if it wasn't one of the
other possibilities, it must be this one.  This is a valid assumption
currently, but future commits will change that.  Prepare for them with
this commit.