perlapi: Consolidate the sv_eq and svstr_eq entries
These functions do the same thing, differing in details.  Consolidation
makes perlapi more compact, and makes it easier for the reader to see
the similarities and differences between the functions.

Note that the cryptic sentence "It correctly handles the UTF8 flag" in
the sv_streq entry has been removed.  It turns out that meant that the
code looks at the UTF8ness of each SV, except when 'use bytes' is in
effect.