Move some ARGS_ASSERT macros to function start
Historically the asserts had to be placed after any declarations
because of limitations in the C89 Standard that have been removed in C99
which we are now following.

Placing the assertions at the function beginning is clearer, and stops
any issues with the code below it using a variable prior to its
assertion.