t/class/accessor.t - add additional :reader and :writer tests
The :reader tests are not really needed at present, as `pp_leavesub` makes
the necessary copies, but exploring some fast accessor ideas which bypass
that logic has shown the usefulness of having these tests present.

A test for :writer on an array field already exists in a separate file.
A comment pointing to that file has been added, and a test for the hash
case added next to the pre-existing test.