S_parse_ident: Restructure croaks
This removes the test for the identifier being too long at the beginning
of the loop.

Almost all the branches already have their own checks for this, and by
restructuring those slightly, this one becomes extraneous, except for
the one branch that didn't have a check.  And it turns out this was
wrong for that branch, which is the one that turns an apostrophe package
separator into ::.  The removed code assumed that the output doesn't
expand, whereas each single apostrophe is in fact replaced by two
characters.  This commit corrects that.

And it consolidates the handling to the target of a goto at the end of
the function.  Right now, there is no real advantage in that, but a
future commit will make the handling more complicated, so a single point
for it will be useful.