ParseXS: refactor: regularise BOOT processing
(This commit is part of a series which will extend the AST parse tree
from just representing individual XSUBs to representing the whole XS
file.)

For some reason most of the file-scoped keywords (PROTOTYPES etc) are
handled together in a single 'while /REQUIRE|PROTOTYPES|.../' loop, but
then BOOT is checked for separately after that loop has completed.

This commit just moves checking for and handling BOOT into that same
loop. The only significant difference between BOOT and the other
keywords is that BOOT consumes all line until the end of the paragraph,
rather than stopping at the next keyword. But this shouldn't make any
difference to the top-level parsing loop.