ParseXS: refactor: remove check_keyword() method
(This commit is part of a series which will extend the AST parse tree
from just representing individual XSUBs to representing the whole XS
file.)

This method used to be used for scanning for keywords and then invoking
old-style FOO_handler()s. Replace its last usage with the new
parse_keywords() method (which looks for keywords and creates
Node::FOO's instead), and remove the method itself.

For now, create a temporary node to accumulate all the file-scoped
keywords in a continuous run, and then call as_code() once at the end.

I.e for the input:

    PROTOTYPES: X
    REQUIRE: Y
    ..

formerly a temporary Node::PROTOTYPES node was created, parsed, and then
as_code() immediately called on it, then freed. Then similarly for
REQUIRE etc.

After this commit, a temporary parent node is created, child nodes are
added to it for each of PROTOTYPES, REQUIRE, etc, and finally as_code()
is called once which calls as_code() for each child.

This is a small step in the process of moving towards deferring calling
as_code() until the whole file has been read in and parsed.

The next commit will make the temporary parent node into something
better.