Change reentr.pl names to be legal
This continues the process started in #23592 to change names with
leading underscores to be legal C.  See that p.r. or
4bb3572f7a1c1f3944b7f58b22b6e7a9ef5faba6 for extensive discussion.

This commit simply removes the leading underscore from each of these
struct members.  Each has file scope given that its parent struct is so,
hence are illegal as-is, but each struct has its own namespace in C, so
these won't collide with anything outside that struct.  (They all are
members of PL_reentrant_buffer.)