ParseXS: refactor: INCLUDE: save state locally
(This commit is part of a series which will extend the AST parse tree
from just representing individual XSUBs to representing the whole XS
file.)

Previously whenever INCLUDE or INCLUDE_COMMAND parsed a new file, the
old state (filename, old lines etc) were saved in a new entry pushed
onto the XS_parse_stack.

The previous commit inlined this push and pull into
Node::include::parse(), which now looked a bit like:

    sub parse() {
        ...
        push @{$pxs->{XS_parse_stack}}, {
            type =>' file',
            Filename => $pxs->{in_filename},
            ...
        }

        ... open new file and parse ...

        my $data = pop @{$pxs->{XS_parse_stack}};
        $pxs->{in_pathname} = $data->{Filepathname};
        ...
    }

The intention over the next few commits is to eliminate XS_parse_stack
completely: the functionality of the stack will be replaced by a nested
series of Node::cpp_scope nodes within the AST. For now, this commit
still pushes/pops an entry (so that #if / #endif scope analysing
continues to work), but saves all the values in a local var. So after
this commit, Node::include::parse() looks like:

    sub parse() {
        ...
        push @{$pxs->{XS_parse_stack}}, { type =>' file' };
        my @saved = ($pxs->{in_filename}, ...);

        ... open new file and parse ...

        pop @{$pxs->{XS_parse_stack}};
        ($pxs->{in_filename}, ...) = @saved;
    }

The rest of this commit is just general tidying up and reordering of
some of the code within parse() without change in functionality. Except
I also added

    $pxs->{line_no} = [];

in addition to the already present

    $pxs->{line}    = [];