toke.c: Remove obsolete assert
This resulted in merging two different pull requests at nearly the same
time.  One put in the assert; the other changed things so it was wrong.
They didn't touch the same lines of code, so rebasing didn't show any
conflict, and if it fails depends on the memory layout, so my testing
didn't fail.