ParseXS: refactor: global_cpp_line: parse CPP line
(This commit is part of a series which will extend the AST parse tree
from just representing individual XSUBs to representing the whole XS
file.)

Make the recently-added Node::global_cpp_line::parse() method
split a C preprocessor line into its components and set some flags
indicating whether its part of a conditional and if so, an if or variant.

Save these values as new fields in the node. They will come in useful
shortly.