Partially revert fd88a92 - it's GvCVGEN(gv), not CvGEN(cv)
This commit partially reverts https://github.com/Perl/perl5/commit/fd88a926b5a38da5bc8cb2bb248b141b48e1c71f

In two places in that commit, setting `GvCVGEN` to zero was changed
to just asserting that the value was already zero, because the CV
had just been created (with a Zeroed body). However, the value is
set on a pre-existing GV, not on the fresh CV, so no assumptions
should be made about the current value of `GvCVGEN`.