ParseXS: fix #elif processing
The XS parser is supposed to be able to do basic processing of
file-scoped C preprocessor directives: in particular, to handle
alternative XSUB declarations; e.g.:

    #if X
    void foo(short i)
    #elif Y
    void foo(long i)
    #else
    void foo(int i)
    #endif

But due to a bug present since this feature was added in perl5.003
(and which no one has complained about for 29 years), the XS parser
was incorrectly looking for the string '#elsif' rather than '#elif'.

So the above would fail, but something using only 'else' would compile
okay:

    #if X
    void foo(short i)
    #else
    void foo(int i)
    #endif

This commit fixes the bug and adds some tests.