ParseXS: refactor: reimplement #if processing
(This commit is part of a series which will extend the AST parse tree
from just representing individual XSUBs to representing the whole XS
file.)

This commit includes a few significant changes to ParseXS which are not
separable into individual commits. Some of these also involve minor
changes in behaviour. Hence a rather long commit message, using
pseudo-POD headers to break things up a bit and give a bit of structure.

=head1 Background

The XS parser does some basic analysis of file-scoped C preprocessor
conditional directives such as #if, #elif, #endif. This is mainly for
the purpose of handling otherwise duplicate XSUBs, e.g.

    #ifdef USE_2ARG

    int foo(int i, int j)

    #else

    int foo(int i)

    #endif

For the above, the parser suppresses any 'duplicate' warning, and adds a
'guard' define at the start of each branch, such as the following in the
emitted C code:

    #ifdef USE_2ARG
    #define XSubPPtmpAAAA 1

    XS_EUPXS(XS_Foo__Bar_foo) { ... }

    #else
    #define XSubPPtmpAAAB 1

    XS_EUPXS(XS_Foo__Bar_foo) { ... }

    #endif

It then uses those guards in the boot code to ensure that only the
correct XSUB(s) (and aliases etc) are registered:

    #if XSubPPtmpAAAA
            newXS_deffile("Foo::Bar::foo", XS_Foo__Bar_foo);
    #endif

    #if XSubPPtmpAAAB
            newXS_deffile("Foo::Bar::foo", XS_Foo__Bar_foo);
    #endif

=head1 The changes in this commit

=head2 Add a Node::cpp_scope node between branches

The XS code above was, before this commit, compiled into an AST
which looked something like:

    XS_file
        preamble
        C_part
        postamble
        cpp_scope[type=main]
            global_cpp_line [if]
            xsub
            global_cpp_line [else]
            xsub
            global_cpp_line [endif]

After this commit, all nodes between each conditional CPP line are
now hung off a new cpp_scope node:

    XS_file
        preamble
        C_part
        postamble
        cpp_scope[type=main]
            global_cpp_line [if]
            cpp_scope[type=if]
                xsub
            global_cpp_line [else]
            cpp_scope[type=if]
                xsub
            global_cpp_line [endif]

=head2 The logic to do the #if/#endif processing is completely
reimplemented from scratch.

Formerly, @{pxs->{XS_parse_stack}} was a stack that was pushed or popped
for every INCLUDE and #if/#endif, and the state about conditionals was
contained in that stack.

After this commit, the state is instead implied by the structure of the
AST; in particular, in the currently active cpp_scope node.
The cpp_scope::parse() method has the fetch_para() loop, and each time
it encounters an #if etc it, it creates a child cpp_scope node and
recursively calls its parse() method. That method returns when it sees
the end of a scope, such as encountering an #elif, #else or #endif or
EOF.

Now, virtually all of the logic involved in maintaining '#if' state is
in cpp_scope::parse(), and all the logic to emit '#define XSubPPtmpAAAB
1' etc code is in cpp_scope::as_code(). This is in contrast to
previously, where the logic and code generation were liberally spread
around in ParseXS.pm and Node.pm

To facilitate this change, two fields have been added to the cpp_scope
node type:

   guard_name - the 'XSubPPtmpAAAB' or whatever to use for this branch;
   seen_xsubs - hash of XSUB names seen in this branch.

After this commit, the data in @{pxs->{XS_parse_stack}}  isn't used;
the next commit will remove it completely.

=head2 Almost all code is now generated after parsing is finished.

At the start of this git branch, code was emitted after each XSUB was
parsed. Now (with a few residual exceptions to be addressed shortly) the
whole file is first parsed into a single AST, and *then* the top-level
as_code() method is called to generate the entire C file from the AST.

Some of the changes in this commit represent the removal of temporary
shims that preserved the old parse/emit ordering, and which are now no
longer required.

=head1 Visible changes in behaviour

Although this commit is mostly intended to be just a refactoring, some
changes in behaviour (in terms of what C code is generated and what
warnings are issued) has occurred.

These are made visible in the changes this commit makes to
t/001-basic.t, which deliberately contains tests for the behaviours
which were about to change.

The changes are:

=head2 fewer warnings

The "Warning: duplicate function definition" warning is now emitted
less often, to avoid false positives. In particular, this fixes GH
19661. That ticket concerned code like

    #if C1
    void foo()
    #endif

    #if C2
    void foo()
    #endif

The logic was such that only one of C1 and C2 could ever be true
(basically DBD:MariaDB was adding a missing DBI function if the
installed DBI was an old one). But the XS parser can't know that.
So it used to warn.

The big change in this commit is to now warn only for two identically
named XSUBs which appear in the *same* cpp_scope. This removes false
positives. It may add false negatives, but that is fairly harmless,
as it just means that bad C code will be generated, and the C compiler
will eventually complain anyway.

The guard definitions are still generated as before.

This change in behaviour was intentional, but was also the easiest: it
meant just checking the seen_xsubs hash for the current cpp_scope node.
For more complex behaviour (if you are considering implementing it), it
may require having the concept of a 'current' cpp_scope, e.g.
$pxs->{cur_cpp_scope}, plus a chain of pointers from the current scope
back to outer ones. (This is slightly problematic, as weak refs aren't
available, as Scalar::Util may not be built yet - so it might require
higher level node(s) to have a destructor which searches for those links
and deletes them first) Once the chain is in place, then it's possible
for a newly-added XSUB to update the parent seen_xsubs hashes too, etc.

=head2 The sequencing of XSubPPtmpAAA, XSubPPtmpAAB etc may change

Previously, the sequence of the guard defines was strictly in the order
they appear in the C file. Now, ones in a nested scope come before outer
ones, e.g.

    #ifdef C1
    #define XSubPPtmpAAB 1
    #  ifdef C2
    #  define XSubPPtmpAAA 1
    ...

This should be harmless.

=head2 XSubPPtmpAAA etc may be placed slightly differently

Previously in something like

    #ifdef C1
    #define X 1
    #define Y 2

the guard would be added after any other CPP lines; the above would
cause this C code to be emitted:

    #ifdef C1
    #define X 1
    #define Y 2
    #define XSubPPtmpAAA 1

but is now:

    #ifdef C1
    #define XSubPPtmpAAA 1
    #define X 1
    #define Y 2

i.e.the guard now always comes *immediately* after the conditional at
the start of the branch it will guard.

This should be harmless.